#1125 Only allow collection hierarchies to be created for undefined strict behavior
Strictness describes how a lookup should perform. Thus, if a request is
strict, the content of a collection should strictly match. A similar
rule is applied for non-strict (loose) requests. This left the question
as to when a collection should be created; when there are no elements?
Instead of making assumptions, this differentiates requests (indicated
by the presence of a strict/loose setting) from registrations (undefined
strict/loose setting).